Access "malicious" fluxmonitor test helper via interface
It was a bit icky to make concreteFluxMonitor public just to expose the test
helper XXXTestingOnlyCreateJob so that the contract could be tested against
aberrant node behavior. This accesses the function via an interface, so at least
the struct is still private, although obviously the method must still be public.

The alternative was to move the test into the fluxmonitor package, but that
created an import cycle.